Migrated to PHP 7.
