+++ .travis.yml
@@ -1,11 +1,7 @@
 language: php
 
 php:
-  - 5.4
-  - 5.5
-  - 5.6
   - 7.0
-  - hhvm
 
 before_install:
   - travis_retry composer self-update
@@ -16,7 +12,4 @@ script:
   - bin/phpunit --verbose
 
 matrix:
-  allow_failures:
-    - php: hhvm
-    - php: 7.0
   fast_finish: true
