+++ composer.json
@@ -5,10 +5,10 @@
     "type": "library",
     "license": "BSD-3-Clause",
     "require": {
-        "php": ">=5.4"
+        "php": ">=7.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "4.5.*"
+        "phpunit/phpunit": "~5.0"
     },
     "config": {
         "bin-dir": "bin"
